+++ .github/workflows/io_tests.yaml
@@ -8,9 +8,11 @@ on:
       - 'ecoscope/io/earthranger.py'
       - 'ecoscope/io/async_earthranger.py'
       - 'ecoscope/io/earthranger_utils.py'
+      - 'ecoscope/io/utils.py'
       - 'tests/test_earthranger_utils.py'
       - 'tests/test_async_earthranger_io.py'
       - 'tests/test_earthranger_io.py'
+      - 'tests/test_io_utils.py'
 
 jobs:
   dev:
