+++ ecoscope/io/earthranger.py
@@ -19,8 +19,9 @@ from ecoscope.io.earthranger_utils import (
     dataframe_to_dict,
     format_iso_time,
     to_gdf,
+    to_hex,
+    pack_columns,
 )
-from ecoscope.io.utils import pack_columns, to_hex
 
 
 class EarthRangerIO(ERClient):
